wip on reports calculation and invoice calculation

+++ database/migrations/2021_10_25_130425_create_taxes_table.php
@@ -15,7 +15,7 @@ class CreateTaxesTable extends Migration
     {
         Schema::create('taxes', function (Blueprint $table) {
             $table->id();
-            $table->float('tax');
+            $table->integer('tax');
             $table->timestamps();
         });
     }
